AGRO_bot.ai-Prototype-main/Agrobot/Agrobot ai final backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pydantic==2.4.2
PyJWT==2.8.0
httpx==0.25.2